/s:symbols using Google Web pages works, for now
While Google seems to have deprecated the old finance web pages used by
the /s: (and related  /div, /day) options, the old pages are still
available at finance.google.com so change the  code to look at that.

But given that even that  page will stop working  or change web content,
it may be better for users to move to /gcsv: option.
That uses the Google Docs Spreadsheet, and the =GOOGLEFINANCE function, so
will be more stable.